Clean up graphQL
