+++ fw/graphql/api.go
@@ -1,0 +1,6 @@
+package graphql
+
+type API interface {
+	GetSchema() string
+	GetResolver() Resolver
+}
