+++ fw/graphql/fake.go
@@ -1,0 +1,1 @@
+package graphql
