+++ fw/graphql/resolver.go
@@ -1,0 +1,3 @@
+package graphql
+
+type Resolver interface{}
