+++ fw/graphql/scalar.go
@@ -1,0 +1,7 @@
+package graphql
+
+type Scalar interface {
+	ImplementsGraphQLType(name string) bool
+	UnmarshalGraphQL(input interface{}) error
+	MarshalJSON() ([]byte, error)
+}
